refactor: Code splitting, improves readability
